Refactoring common logic out to a shared class
